Create SECURITY.md
Refer to my writing about people not being trusted.  I can think of a very small handful of reasons that necessitate designation specifically stating "white check mark" and they are not things that showcase good qualities of people and if i am wrong, then I'm wrong, but I'm not.  The risks, imvho, are not something that should be allowed because in these kinds of situations, man is usually categorizing things like "white mark" for bad things.  I would also suggest cross correlation of the words "white mark" with "blank mark" or "blank box" to verify if this is about payment or for vile reasons.